+++ lib/src/llamadart_chat_options.dart
@@ -29,7 +29,6 @@ class LlamadartChatOptions extends ChatModelOptions {
     this.repeatPenalty,
   });
 
-  @override
   LlamadartChatOptions copyWith({
     int? nCtx,
     int? nGpuLayers,